requirements.txt
streamlit>=1.31.0
streamlit-webrtc>=0.47.1
requests>=2.31.0
python-dotenv>=1.0.0
streamlit-lottie>=0.0.5